refactor(package): NoTripleRepeat 유효성 검사 클래스를 config.validator 패키지로 이동
