+++ src/main/java/org/example/goormssd/usermanagementbackend/dto/request/LoginRequestDto.java
@@ -9,7 +9,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.example.goormssd.usermanagementbackend.util.NoTripleRepeat;
+import org.example.goormssd.usermanagementbackend.config.validator.NoTripleRepeat;
 
 @Getter
 @Setter
